+++ src/message/index.ts
@@ -146,3 +146,15 @@ export {
 
 // Control Message Union
 export type { ControlMessage } from "./control";
+
+// AUTHORIZATION_TOKEN Token 構造
+export {
+  type AuthToken,
+  type AuthTokenDelete,
+  type AuthTokenRegister,
+  type AuthTokenUseAlias,
+  type AuthTokenUseValue,
+  AuthTokenAliasType,
+  decodeAuthToken,
+  encodeAuthToken,
+} from "./authToken";